requirements_win32.txt
androidviewclient==23.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
culebratester-client==2.0.64
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
et-xmlfile==1.1.0
fonttools==4.46.0
idna==3.6
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
numpy==1.26.2
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.36
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
questionary==2.0.1
requests==2.31.0
rich==13.7.0
six==1.16.0
https://github.com/simonflueckiger/tesserocr-windows_build/releases/download/tesserocr-v2.6.0-tesseract-5.3.1/tesserocr-2.6.0-cp311-cp311-win32.whl
urllib3==2.1.0
wcwidth==0.2.12
